# Flask-mysql-python/src/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Productos y Pedidos",
    "description": "API para gestionar productos y pedidos",
    "version": "1.0.0"
  },
  "paths": {
    "/api/productos": {
      "get": {
        "summary": "Obtener todos los productos",
        "responses": {
          "200": {
            "description": "Lista de productos",
            "content": {
              "application/json": {
                "example": {
                  "id": 1,
                  "nombre": "Producto 1",
                  "descripcion": "Descripción del producto 1",
                  "categoria": "Categoría 1",
                  "precio": 10.99,
                  "stock": 100
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear un nuevo producto",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string"
                  },
                  "descripcion": {
                    "type": "string"
                  },
                  "categoria": {
                    "type": "string"
                  },
                  "precio": {
                    "type": "number"
                  },
                  "stock": {
                    "type": "integer"
                  }
                },
                "required": ["nombre", "descripcion", "categoria", "precio", "stock"]
              },
              "example": {
                "nombre": "Nuevo Producto",
                "descripcion": "Descripción del nuevo producto",
                "categoria": "Categoría del nuevo producto",
                "precio": 19.99,
                "stock": 50
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Producto creado exitosamente"
          }
        }
      }
    },
    "/api/productos/{id}": {
      "get": {
        "summary": "Obtener un producto específico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID del producto",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles del producto",
            "content": {
              "application/json": {
                "example": {
                  "id": 1,
                  "nombre": "Producto 1",
                  "descripcion": "Descripción del producto 1",
                  "categoria": "Categoría 1",
                  "precio": 10.99,
                  "stock": 100
                }
              }
            }
          },
          "404": {
            "description": "Producto no encontrado"
          }
        }
      },
      "put": {
        "summary": "Actualizar un producto",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID del producto",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string"
                  },
                  "descripcion": {
                    "type": "string"
                  },
                  "categoria": {
                    "type": "string"
                  },
                  "precio": {
                    "type": "number"
                  },
                  "stock": {
                    "type": "integer"
                  }
                },
                "required": ["nombre", "descripcion", "categoria", "precio", "stock"]
              },
              "example": {
                "nombre": "Producto Actualizado",
                "descripcion": "Descripción del producto actualizado",
                "categoria": "Categoría del producto actualizado",
                "precio": 24.99,
                "stock": 75
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Producto actualizado exitosamente"
          },
          "404": {
            "description": "Producto no encontrado"
          }
        }
      },
      "delete": {
        "summary": "Eliminar un producto",
        "description": "Elimina un producto por su ID.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "ID del producto a eliminar",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Producto eliminado exitosamente"
          },
          "404": {
            "description": "Producto no encontrado"
          }
        },
        "x-code-samples": [
          {
            "lang": "Python",
            "source": "cursor = db.database.cursor()\n\n# Eliminar los registros de la tabla pedido_productos\ncursor.execute(\"DELETE FROM pedido_productos WHERE producto_id = %s\", (id,))\n\n# Eliminar el producto\ncursor.execute(\"DELETE FROM producto WHERE id = %s\", (id,))\n\n# Confirmar los cambios\ndb.database.commit()\n\ncursor.close()\n\nreturn make_response(jsonify({'message': 'Producto eliminado exitosamente'}), 200)"
          }
        ]
      }
    },
    "/api/pedidos": {
      "get": {
        "summary": "Obtener todos los pedidos",
        "responses": {
          "200": {
            "description": "Lista de pedidos",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": 1,
                    "rut_cliente": "12345678-9",
                    "fecha_pedido": "2023-05-01",
                    "productos": [
                      {
                        "id": 1,
                        "nombre": "Producto 1",
                        "cantidad": 2
                      },
                      {
                        "id": 2,
                        "nombre": "Producto 2",
                        "cantidad": 1
                      }
                    ]
                  },
                  {
                    "id": 2,
                    "rut_cliente": "98765432-1",
                    "fecha_pedido": "2023-05-05",
                    "productos": [
                      {
                        "id": 3,
                        "nombre": "Producto 3",
                        "cantidad": 3
                      }
                    ]
                  }
                ]
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear un nuevo pedido",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rut_cliente": {
                    "type": "string"
                  },
                  "fecha_pedido": {
                    "type": "string",
                    "format": "date"
                  },
                  "productos": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "producto_id": {
                          "type": "integer"
                        },
                        "cantidad": {
                          "type": "integer"
                        }
                      },
                      "required": ["producto_id", "cantidad"]
                    }
                  }
                },
                "required": ["rut_cliente", "fecha_pedido", "productos"]
              },
              "example": {
                "rut_cliente": "12345678-9",
                "fecha_pedido": "2023-05-10",
                "productos": [
                  {
                    "producto_id": 1,
                    "cantidad": 2
                  },
                  {
                    "producto_id": 2,
                    "cantidad": 1
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pedido creado exitosamente"
          }
        }
      }
    },
    "/api/pedidos/{pedido_id}": {
      "get": {
        "summary": "Obtener un pedido específico",
        "parameters": [
          {
            "name": "pedido_id",
            "in": "path",
            "required": true,
            "description": "ID del pedido",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pedido encontrado",
            "content": {
              "application/json": {
                "example": {
                  "id": 1,
                  "rut_cliente": "12345678-9",
                  "fecha_pedido": "2023-05-01",
                  "productos": [
                    {
                      "id": 1,
                      "nombre": "Producto 1",
                      "cantidad": 2
                    },
                    {
                      "id": 2,
                      "nombre": "Producto 2",
                      "cantidad": 1
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Pedido no encontrado"
          }
        }
      },
      "put": {
        "summary": "Actualizar un pedido",
        "parameters": [
          {
            "name": "pedido_id",
            "in": "path",
            "required": true,
            "description": "ID del pedido",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rut_cliente": {
                    "type": "string"
                  },
                  "fecha_pedido": {
                    "type": "string",
                    "format": "date"
                  },
                  "productos": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "producto_id": {
                          "type": "integer"
                        },
                        "cantidad": {
                          "type": "integer"
                        }
                      },
                      "required": ["producto_id", "cantidad"]
                    }
                  }
                },
                "required": ["rut_cliente", "fecha_pedido", "productos"]
              },
              "example": {
                "rut_cliente": "12345678-9",
                "fecha_pedido": "2023-05-15",
                "productos": [
                  {
                    "producto_id": 1,
                    "cantidad": 3
                  },
                  {
                    "producto_id": 2,
                    "cantidad": 2
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pedido actualizado exitosamente"
          },
          "404": {
            "description": "Pedido no encontrado"
          }
        }
      },
      "delete": {
        "summary": "Eliminar un pedido",
        "parameters": [
          {
            "name": "pedido_id",
            "in": "path",
            "required": true,
            "description": "ID del pedido",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pedido eliminado exitosamente"
          },
          "404": {
            "description": "Pedido no encontrado"
          }
        }
      }
    }
  }
}